carrier-predictor/requirements.txt
# Core API
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1

# ML & Embeddings
sentence-transformers==3.3.1
faiss-cpu==1.9.0
torch==2.5.1
numpy==2.1.3

# Document Processing
pypdf==5.1.0
trafilatura==1.12.2

# YAML Support
pyyaml==6.0.2

# Optional: OpenAI
openai==1.55.3

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.2

# Code Quality (optional)
black==24.10.0
isort==5.13.2
ruff==0.8.2
